pyproject.toml
[project]
name = "agent-samples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "adktools>=0.1.2",
    "dotenv>=0.9.9",
    "google-adk>=1.1.1",
    "litellm>=1.72.2",
    "pydantic>=2.11.5",
]
